requirements.txt
nltk==3.4.4
google-cloud-language==1.2.0
